# Speedup logging by 28%