--- a/pyftpdlib/log.py
+++ b/pyftpdlib/log.py
@@ -149,12 +149,12 @@ def is_logging_configured():
 # TODO: write tests
 
 def config_logging(level=LEVEL, prefix=PREFIX, other_loggers=None):
-    # Speedup logging by preventing certain log record info to be
-    # fetched, see:
+    # Speedup logging by preventing certain internal log record info to
+    # be unnecessarily fetched, see:
     # https://docs.python.org/3/howto/logging.html#optimization
-    # https://docs.python.org/3/library/logging.html#logrecord-attributes
-    # https://stackoverflow.com/a/38924153/376587
-    # This results in about 28% speedup.
+    # https://docs.python.org/3/library/logging.html#logrecord-
+    # attributes https://stackoverflow.com/a/38924153/376587 This
+    # results in about 28% speedup.
     key_names = re.findall(
         r'(?<!%)%\(([^)]+)\)[-# +0-9.hlL]*[diouxXeEfFgGcrs]', prefix)
     if "process" not in key_names:
@@ -167,6 +167,7 @@ def config_logging(level=LEVEL, prefix=PREFIX, other_loggers=None):
             "pathname" not in key_names and \
             "lineno" not in key_names and \
             "module" not in key_names:
+        # biggest speedup as it avoids calling sys._getframe()
         logging._srcfile = None
     handler = logging.StreamHandler()
     formatter = LogFormatter()